Move displayconfig utils to lib
