+++ src/services/useDisplayConfig/index.ts
@@ -1,23 +1,22 @@
 import {
-  ActionResult,
   PiWebserviceProvider,
-  TimeSeriesDisplaySubplot,
   type ActionsResponse,
 } from '@deltares/fews-pi-requests'
 import { computed, ref, toValue, watch, watchEffect } from 'vue'
 import type { MaybeRefOrGetter, Ref } from 'vue'
-import { DisplayConfig } from '../../lib/display/DisplayConfig.js'
-import { timeSeriesDisplayToChartConfig } from '../../lib/charts/timeSeriesDisplayToChartConfig.js'
+import type { DisplayConfig } from '@/lib/display'
 import { createTransformRequestFn } from '@/lib/requests/transformRequest.js'
-import { MD5 } from 'crypto-js'
-import { convertFewsPiDateTimeToJsDate } from '@/lib/date'
 import {
   type Filter,
   isFilterActionsFilter,
   isTimeSeriesGridActionsFilter,
 } from '@/lib/filters'
 import { useTaskRunColorsStore } from '@/stores/taskRunColors'
-import { addFilterPeriodToConfig } from '@/lib/display'
+import {
+  actionsResponseToDisplayConfig,
+  addFilterPeriodToConfig,
+  addIndexToKeys,
+} from '@/lib/display'
 
 export interface UseDisplayConfigReturn {
   displayConfig: Ref<DisplayConfig | null>
@@ -27,79 +26,6 @@ export interface UseDisplayConfigReturn {
 export interface UseDisplayConfigOptions {
   convertDatum?: boolean
   useDisplayUnits?: boolean
-}
-
-/**
- * Applies colors to subplot items based on their taskRunIds.
- * @param subPlot The subplot to which colors will be applied.
- */
-function applyColorsToSubplotItems(subPlot: TimeSeriesDisplaySubplot): void {
-  const taskRunColorsStore = useTaskRunColorsStore()
-  for (const item of subPlot.items) {
-    // find the taskRunId in the actionRequest array
-    const taskRunId = item.taskRunId
-    if (!taskRunId) continue
-    // Get color from the taskRunColorsStore
-    const color = taskRunColorsStore.getColor(taskRunId)
-    if (color) {
-      item.color = color
-    }
-  }
-}
-
-/**
- * Converts the actions response to an array of display configurations.
- * @param actionsResponse The actions response object.
- * @returns An array of display configurations.
- */
-function actionsResponseToDisplayConfig(
-  actionsResponse: ActionsResponse,
-  nodeId: string | undefined,
-  startTime?: Date,
-  endTime?: Date,
-): DisplayConfig[] {
-  const displays: DisplayConfig[] = []
-  for (const result of actionsResponse.results) {
-    if (result.config === undefined) continue
-    const title = result.config.timeSeriesDisplay.title ?? ''
-    const timeSeriesDisplayIndex = result.config.timeSeriesDisplay.index
-    const period = result.config.timeSeriesDisplay.period
-    const plotId = result.config.timeSeriesDisplay.plotId
-
-    // The period is always specified in UTC.
-    const timeZoneOffsetString = 'Z'
-    let configPeriod: [Date, Date]
-    if (period) {
-      const periodStart = convertFewsPiDateTimeToJsDate(
-        period.startDate,
-        timeZoneOffsetString,
-      )
-      const periodEnd = convertFewsPiDateTimeToJsDate(
-        period.endDate,
-        timeZoneOffsetString,
-      )
-      configPeriod = [startTime ?? periodStart, endTime ?? periodEnd]
-    }
-    const subplots =
-      result.config.timeSeriesDisplay.subplots?.map((subPlot) => {
-        applyColorsToSubplotItems(subPlot)
-        return timeSeriesDisplayToChartConfig(subPlot, configPeriod)
-      }) ?? []
-    const display: DisplayConfig = {
-      id: title,
-      title,
-      forecastLegend: result.config.timeSeriesDisplay.forecastLegend,
-      plotId,
-      nodeId: nodeId,
-      class: 'singles',
-      index: timeSeriesDisplayIndex,
-      requests: result.requests,
-      period: result.config.timeSeriesDisplay.period,
-      subplots,
-    }
-    displays.push(display)
-  }
-  return displays
 }
 
 /**
@@ -283,23 +209,4 @@ export function useDisplayConfigFilter(
   }
 
   return shell
-}
-
-function addIndexToKeys(results: ActionResult[]) {
-  results.forEach((result) => {
-    result.requests.forEach((request) => {
-      request.key = MD5(request.request).toString()
-    })
-    if (result.config?.timeSeriesDisplay.subplots) {
-      let i = 0
-      result.config.timeSeriesDisplay.subplots.forEach((subPlot) => {
-        subPlot.items.forEach((item) => {
-          if (item.request === undefined) {
-            item.request = `${result.requests[0].key}[${i}]`
-          }
-          i++
-        })
-      })
-    }
-  })
 }
